# src/lib/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Calendar & location hub API",
    "version": "1.0.0",
    "description": "Schedule, sources, locations, sharing, and manual calendar items.\n\n**Authentication:** Most routes require an HTTP-only session cookie `cal_owner` after `POST /api/auth/login` (or dev mode with `APP_PASSWORD` unset). Use **Try it out** in the browser while logged in so cookies are sent.\n\n**Public:** `GET /api/plan/{token}/schedule`, auth login/logout/config, OAuth callbacks/start, and this OpenAPI document."
  },
...
  "servers": [{ "url": "http://localhost:3000", "description": "Overridden at runtime by /api/openapi" }],
  "components": {
    "securitySchemes": {
      "OwnerCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cal_owner",
        "description": "JWT set by POST /api/auth/login"
      }
    },
    "schemas": {
...
      "SchedulePayload": {
        "type": "object",
        "properties": {
          "rangeStart": { "type": "string", "format": "date-time" },
          "rangeEnd": { "type": "string", "format": "date-time" },
          "slots": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "startUtc": { "type": "string" },
                "endUtc": { "type": "string" },
                "busy": { "type": "boolean" },
                "eventTitle": { "type": "string", "nullable": true },
                "eventLocation": { "type": "string", "nullable": true },
                "sourceName": { "type": "string", "nullable": true },
                "sourceColor": { "type": "string", "nullable": true },
                "placeLabel": { "type": "string", "nullable": true },
                "placeDetail": { "type": "string", "nullable": true }
              }
            }
          },
          "events": { "type": "array", "items": { "type": "object", "additionalProperties": true } },
          "locationWindows": { "type": "array", "items": { "type": "object", "additionalProperties": true } }
        }
      },
      "CalendarSource": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "kind": { "type": "string" },
          "name": { "type": "string" },
          "icalUrl": { "type": "string", "nullable": true },
          "color": { "type": "string" },
          "eventCount": { "type": "integer" },
          "accountHint": { "type": "string", "nullable": true },
          "calendarId": { "type": "string", "nullable": true }
        }
      },
...
    }
  },
  "paths": {
...
    "/api/schedule": {
      "get": {
        "tags": ["Schedule"],
        "summary": "Owner schedule for a time range",
        "parameters": [
          { "name": "from", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Range start (default: today local)" },
          { "name": "to", "in": "query", "schema": { "type": "string", "format": "date-time" }, "description": "Range end (default: from + 14 days)" }
        ],
        "responses": {
          "200": { "description": "Merged schedule", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SchedulePayload" } } } },
          "400": { "description": "Invalid range" },
          "401": { "description": "Unauthorized" },
          "500": { "description": "Server error" }
        },
        "security": [{ "OwnerCookie": [] }]
      }
    },
...
    "/api/sources": {
      "get": {
        "tags": ["Sources"],
        "summary": "List calendar sources",
        "responses": {
          "200": {
            "content": {
              "application/json": { "schema": { "type": "array", "items": { "$ref": "#/components/schemas/CalendarSource" } } }
            }
          },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "OwnerCookie": [] }]
      },
      "post": {
        "tags": ["Sources"],
        "summary": "Add ICS (webcal) source only; OAuth uses /api/auth/.../start",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["kind", "name", "icalUrl"],
                "properties": {
                  "kind": { "type": "string", "enum": ["ICAL"] },
                  "name": { "type": "string" },
                  "icalUrl": { "type": "string", "format": "uri" },
                  "color": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CalendarSource" } } } },
          "400": { "description": "Validation error" },
          "401": { "description": "Unauthorized" }
        },
        "security": [{ "OwnerCookie": [] }]
      }
    },
...
  }
}